requirements.txt
django==4.2.14
djangorestframework==3.15.2
celery==5.4
redis==5.0.7
eventlet==0.36.1